+++ lib/fugit.rb
@@ -3,7 +3,7 @@
 
 module Fugit
 
-  VERSION = '1.12.0'
+  VERSION = '1.12.1'
 end
 
 require 'time'