+++ lib/fugit/cron.rb
@@ -834,7 +834,8 @@ module Fugit
           else
             a.collect(&:to_s).join('#')
           end }
-        .join(',')
+        .join(',') +
+      (@day_and ? '&' : '')
     end
 
     WEEKDAYS =